Started basic restructuring of app
