+++ config.ru
@@ -1,0 +1,3 @@
+root = ::File.dirname(__FILE__)
+require ::File.join( root, 'hashview' )
+run HashView.new
