+++ helpers/hc_stdout_parser.rb
@@ -1,25 +1,28 @@
-# this helper parses the hashcat stdout
+# encoding: utf-8
 
-def hashcatParser(file)
-  status = {}
-  File.open(file).each_line do |line|
-    if line.start_with?('Time.Started.')
-      status['Time_Started'] = line.split(': ')[-1].strip
-    elsif line.start_with?('Time.Estimated.')
-      status['Time_Estimated'] = line.split(': ')[-1].strip
-    elsif line.start_with?('Recovered.')
-      status['Recovered'] = line.split(': ')[-1].strip
-    elsif line.start_with?('Input.Mode.')
-      status['Input_Mode'] = line.split(': ')[-1].strip
-    elsif line.start_with?('Speed.Dev.')
-      item = line.split(': ')
-      gpu = item[0].gsub!('Speed.Dev.', 'Speed Dev ').gsub!('.', '')
-      status[gpu] = line.split(': ')[-1].strip
-    elsif line.start_with?('HWMon.Dev.')
-      item = line.split('.: ')
-      gpu = item[0].gsub!('HWMon.Dev.', 'HWMon Dev ').gsub!('.', '')
-      status[gpu] = line.split('.: ')[-1].strip
+# this helper parses the hashcat stdout
+module HcStdoutParser 
+  def hashcatParser(file)
+    status = {}
+    File.open(file).each_line do |line|
+      if line.start_with?('Time.Started.')
+        status['Time_Started'] = line.split(': ')[-1].strip
+      elsif line.start_with?('Time.Estimated.')
+        status['Time_Estimated'] = line.split(': ')[-1].strip
+      elsif line.start_with?('Recovered.')
+        status['Recovered'] = line.split(': ')[-1].strip
+      elsif line.start_with?('Input.Mode.')
+        status['Input_Mode'] = line.split(': ')[-1].strip
+      elsif line.start_with?('Speed.Dev.')
+        item = line.split(': ')
+        gpu = item[0].gsub!('Speed.Dev.', 'Speed Dev ').gsub!('.', '')
+        status[gpu] = line.split(': ')[-1].strip
+      elsif line.start_with?('HWMon.Dev.')
+        item = line.split('.: ')
+        gpu = item[0].gsub!('HWMon.Dev.', 'HWMon Dev ').gsub!('.', '')
+        status[gpu] = line.split('.: ')[-1].strip
+      end
     end
+    return status
   end
-  return status
 end
