+++ helpers/sintra_ssl.rb
@@ -1,3 +1,4 @@
+# encoding utf-8
 require 'webrick/https'
 
 module Sinatra
