+++ main.rb
@@ -231,539 +231,19 @@ end
 
 ### customer controllers ###
 
-get '/customers/list' do
-  @customers = Customers.all(order: [:name.asc])
-  @total_jobs = []
-  @total_hashfiles = []
-
-  @customers.each do |customer|
-    @total_jobs[customer.id] = Jobs.count(customer_id: customer.id)
-    @total_hashfiles[customer.id] = Hashfiles.count(customer_id: customer.id)
-  end
-
-  haml :customer_list
-end
-
-get '/customers/create' do
-  haml :customer_edit
-end
-
-post '/customers/create' do
-  varWash(params)
-
-  if !params[:name] || params[:name].nil?
-    flash[:error] = 'Customer must have a name.'
-    redirect to('/customers/create')
-  end
-
-  customer = Customers.new
-  customer.name = params[:name]
-  customer.description = params[:desc]
-  customer.save
-
-  redirect to('customers/list')
-end
-
-get '/customers/edit/:id' do
-  varWash(params)
-  @customer = Customers.first(id: params[:id])
-
-  haml :customer_edit
-end
-
-post '/customers/edit/:id' do
-  varWash(params)
-  if !params[:name] || params[:name].nil?
-    flash[:error] = 'Customer must have a name.'
-    redirect to('/customers/create')
-  end
-
-  customer = Customers.first(id: params[:id])
-  customer.name = params[:name]
-  customer.description = params[:desc]
-  customer.save
-
-  redirect to('customers/list')
-end
-
-get '/customers/delete/:id' do
-  varWash(params)
-
-  @customer = Customers.first(id: params[:id])
-  @customer.destroy unless @customer.nil?
-
-  @jobs = Jobs.all(customer_id: params[:id])
-  unless @jobs.nil?
-    @jobs.each do |job|
-      @jobtasks = Jobtasks.all(job_id: job.id)
-      @jobtasks.destroy unless @jobtasks.nil?
-    end
-    @jobs.destroy unless @jobs.nil?
-  end
-
-  # @hashfilehashes = Hashfilehashes.all(hashfile_id:
-  # Need to select/identify what hashfiles are associated with this customer then remove them from hashfilehashes 
-
-  @hashfiles = Hashfiles.all(customer_id: params[:id])
-  @hashfiles.destroy unless @hashfiles.nil?
-
-  redirect to('/customers/list')
-end
-
-post '/customers/upload/hashfile' do
-  varWash(params)
-
-  if params[:hashfile_name].nil? || params[:hashfile_name].empty?
-    flash[:error] = 'You must specificy a name for this hash file.'
-    redirect to("/jobs/assign_hashfile?customer_id=#{params[:customer_id]}&job_id=#{params[:job_id]}")
-  end
-
-  if params[:file].nil? || params[:file].empty?
-    flash[:error] = 'You must specify a hashfile.'
-    redirect to("/jobs/assign_hashfile?customer_id=#{params[:customer_id]}&job_id=#{params[:job_id]}")
-  end
-
-  @job = Jobs.first(id: params[:job_id])
-  return 'No such job exists' unless @job
-
-  # temporarily save file for testing
-  hash = rand(36**8).to_s(36)
-  hashfile = "control/hashes/hashfile_upload_job_id-#{@job.id}-#{hash}.txt"
-
-  # Parse uploaded file into an array
-  hash_array = []
-  whole_file_as_string_object = params[:file][:tempfile].read
-  File.open(hashfile, 'w') { |f| f.write(whole_file_as_string_object) }
-  whole_file_as_string_object.each_line do |line|
-    hash_array << line
-  end
-
-  # save location of tmp hash file
-  hashfile = Hashfiles.new
-  hashfile.name = params[:hashfile_name]
-  hashfile.customer_id = params[:customer_id]
-  hashfile.hash_str = hash
-  hashfile.save
-
-  @job.save
-
-  redirect to("/customers/upload/verify_filetype?customer_id=#{params[:customer_id]}&job_id=#{params[:job_id]}&hashid=#{hashfile.id}")
-end
-
-get '/customers/upload/verify_filetype' do
-  varWash(params)
-
-  hashfile = Hashfiles.first(id: params[:hashid])
-
-  @filetypes = detectHashfileType("control/hashes/hashfile_upload_job_id-#{params[:job_id]}-#{hashfile.hash_str}.txt")
-  @job = Jobs.first(id: params[:job_id])
-  haml :verify_filetypes
-end
-
-post '/customers/upload/verify_filetype' do
-  varWash(params)
-
-  redirect to("/customers/upload/verify_hashtype?customer_id=#{params[:customer_id]}&job_id=#{params[:job_id]}&hashid=#{params[:hashid]}&filetype=#{params[:filetype]}")
-end
-
-get '/customers/upload/verify_hashtype' do
-  varWash(params)
-
-  hashfile = Hashfiles.first(id: params[:hashid])
-
-  @hashtypes = detectHashType("control/hashes/hashfile_upload_job_id-#{params[:job_id]}-#{hashfile.hash_str}.txt", params[:filetype])
-  @job = Jobs.first(id: params[:job_id])
-  haml :verify_hashtypes
-end
-
-post '/customers/upload/verify_hashtype' do
-  varWash(params)
-
-  if !params[:filetype] || params[:filetype].nil?
-    flash[:error] = 'You must specify a valid hashfile type.'
-    redirect to("/customers/upload/verify_hashtype?customer_id=#{params[:customer_id]}&job_id=#{params[:job_id]}&hashid=#{params[:hashid]}&filetype=#{params[:filetype]}")
-  end
-
-  filetype = params[:filetype]
-
-  hashfile = Hashfiles.first(id: params[:hashid])
-
-  if params[:hashtype] == '99999'
-    hashtype = params[:manualHash]
-  else
-    hashtype = params[:hashtype]
-  end
-
-  hash_file = "control/hashes/hashfile_upload_job_id-#{params[:job_id]}-#{hashfile.hash_str}.txt"
-
-  hash_array = []
-  File.open(hash_file, 'r').each do |line|
-    hash_array << line
-  end
-
-  @job = Jobs.first(id: params[:job_id])
-  @job.hashfile_id = hashfile.id
-  @job.save
-
-  unless importHash(hash_array, hashfile.id, filetype, hashtype)
-    flash[:error] = 'Error importing hashes'
-    redirect to("/customers/upload/verify_hashtype?customer_id=#{params[:customer_id]}&job_id=#{params[:job_id]}&hashid=#{params[:hashid]}&filetype=#{params[:filetype]}")
-  end
-
-  previously_cracked_cnt = repository(:default).adapter.select('SELECT COUNT(h.originalhash) FROM hashes h LEFT JOIN hashfilehashes a ON h.id = a.hash_id WHERE (a.hashfile_id = ? AND h.cracked = 1)', hashfile.id)[0].to_s
-  total_cnt = repository(:default).adapter.select('SELECT COUNT(h.originalhash) FROM hashes h LEFT JOIN hashfilehashes a ON h.id = a.hash_id WHERE a.hashfile_id = ?', hashfile.id)[0].to_s
-
-  unless total_cnt.nil?
-    flash[:success] = 'Successfully uploaded ' + total_cnt + ' hashes.'
-  end
-
-  unless previously_cracked_cnt.nil?
-    flash[:success] = previously_cracked_cnt + ' have already been cracked!'
-  end
-
-  # Delete file, no longer needed
-  File.delete(hash_file)
-
-  if params[:edit]
-    redirect to("/jobs/assign_tasks?job_id=#{params[:job_id]}&edit=1")
-  else
-    redirect to("/jobs/assign_tasks?job_id=#{params[:job_id]}")
-  end
-end
+## Moved to routes/customers.rb
 
 ############################
 
 ### Account controllers ####
 
-get '/accounts/list' do
-  @users = User.all
-
-  haml :account_list
-end
-
-get '/accounts/create' do
-  haml :account_edit
-end
-
-post '/accounts/create' do
-  varWash(params)
-
-  if params[:username].nil? || params[:username].empty?
-    flash[:error] = 'You must have username.'
-    redirect to('/accounts/create')
-  end
-
-  if params[:password].nil? || params[:password].empty?
-    flash[:error] = 'You must have a password.'
-    redirect to('/accounts/create')
-  end
-
-  if params[:confirm].nil? || params[:confirm].empty?
-    flash[:error] = 'You must have a password.'
-    redirect to('/accounts/create')
-  end
-
-  # validate that no other user account exists
-  @users = User.all(username: params[:username])
-  if @users.empty?
-    if params[:password] != params[:confirm]
-      flash[:error] = 'Passwords do not match'
-      redirect to('/accounts/create')
-    else
-      new_user = User.new
-      new_user.username = params[:username]
-      new_user.password = params[:password]
-      new_user.email = params[:email] unless params[:email].nil? || params[:email].empty?
-      new_user.admin = 't'
-      new_user.save
-    end
-  else
-    flash[:error] = 'User account already exists.'
-    redirect to('/accounts/create')
-  end
-  redirect to('/accounts/list')
-end
-
-get '/accounts/edit/:account_id' do
-  varWash(params)
-
-  @user = User.first(id: params[:account_id])
-
-  haml :account_edit
-end
-
-post '/accounts/save' do
-  varWash(params)
-
-  if params[:account_id].nil? || params[:account_id].empty?
-    flash[:error] = 'Invalid account.'
-    redirect to('/accounts/list')
-  end
-
-  if params[:username].nil? || params[:username].empty?
-    flash[:error] = 'Invalid username.'
-    redirect to("/accounts/edit/#{params[:account_id]}")
-  end
-
-  if params[:password] != params[:confirm]
-    flash[:error] = 'Passwords do not match.'
-    redirect to("/accounts/edit/#{params[:account_id]}")
-  end
-
-  user = User.first(id: params[:account_id])
-  user.username = params[:username]
-  user.password = params[:password] unless params[:password].nil? || params[:password].empty?
-  user.email = params[:email] unless params[:email].nil? || params[:email].empty?
-  user.save
-
-  flash[:success] = 'Account successfuly updated.'
-
-  redirect to('/accounts/list')
-end
-
-get '/accounts/delete/:id' do
-  varWash(params)
-
-  @user = User.first(id: params[:id])
-  @user.destroy unless @user.nil?
-
-  redirect to('/accounts/list')
-end
+## Moved to routes/accounts.rb
 
 ############################
 
 ##### task controllers #####
 
-get '/tasks/list' do
-  @tasks = Tasks.all
-  @wordlists = Wordlists.all
-
-  haml :task_list
-end
-
-get '/tasks/delete/:id' do
-  varWash(params)
-
-  @job_tasks = Jobtasks.all(task_id: params[:id])
-  unless @job_tasks.empty?
-    flash[:error] = 'That task is currently used in a job.'
-    redirect to('/tasks/list')
-  end
-
-  @task = Tasks.first(id: params[:id])
-  @task.destroy if @task
-
-  redirect to('/tasks/list')
-end
-
-get '/tasks/edit/:id' do
-  varWash(params)
-  @task = Tasks.first(id: params[:id])
-  @wordlists = Wordlists.all
-  @settings = Settings.first
-
-  if @task.hc_attackmode == 'combinator'
-    @combinator_wordlists = @task.wl_id.split(',')
-    if @task.hc_rule =~ /--rule-left=(.*) --rule-right=(.*)/
-      @combinator_left_rule = $1
-      @combinator_right_rule = $2
-    elsif @task.hc_rule =~ /--rule-left=(.*)/
-      @combinator_left_rule = $1
-    elsif @task.hc_rule =~ /--rule-right=(.*)/
-      @combinator_right_rule = $1
-    end
-  end
-
-  @rules = []
-  # list wordlists that can be used
-  Dir.foreach('control/rules/') do |item|
-    next if item == '.' || item == '..'
-      @rules << item
-  end
-
-  haml :task_edit
-end
-
-post '/tasks/edit/:id' do
-  varWash(params)
-  if !params[:name] || params[:name].nil?
-    flash[:error] = 'The task requires a name.'
-    redirect to("/tasks/edit/#{params[:id]}")
-  end
-
-  settings = Settings.first
-  wordlist = Wordlists.first(id: params[:wordlist])
-
-  if settings && !settings.hcbinpath
-    flash[:error] = 'No hashcat binary path is defined in global settings.'
-    redirect to('/settings')
-  end
-
-  # must have two word lists
-  if params[:attackmode] == 'combinator'
-    wordlist_count = 0
-    wordlist_list = ''
-    rule_list = ''
-    @wordlists = Wordlists.all
-    @wordlists.each do |wordlist_check|
-      params.keys.each do |key|
-        if params[key] == 'on' && key == "combinator_wordlist_#{wordlist_check.id}"
-          if wordlist_list == ''
-            wordlist_list = wordlist_check.id.to_s + ','
-          else
-            wordlist_list = wordlist_list + wordlist_check.id.to_s
-          end
-          wordlist_count = wordlist_count + 1
-        end
-      end
-    end
-
-    if wordlist_count != 2
-      flash[:error] = 'You must specify at exactly 2 wordlists.'
-      redirect to("/tasks/edit/#{params[:id]}")
-    end
-
-    if params[:combinator_left_rule] && !params[:combinator_left_rule].empty? && params[:combinator_right_rule] && !params[:combinator_right_rule].empty?
-      rule_list = '--rule-left=' + params[:combinator_left_rule] + ' --rule-right=' + params[:combinator_right_rule]
-    elsif params[:combinator_left_rule] && !params[:combinator_left_rule].empty?
-      rule_list = '--rule-left=' + params[:combinator_left_rule]
-    elsif params[:combinator_right_rule] && !params[:combinator_right_rule].empty?
-      rule_list = '--rule-right=' + params[:combinator_right_rule]
-    else
-      rule_list = ''
-    end
-  end
-
-  task = Tasks.first(id: params[:id])
-  task.name = params[:name]
-
-  task.hc_attackmode = params[:attackmode]
-
-  if params[:attackmode] == 'dictionary'
-    task.wl_id = wordlist.id
-    task.hc_rule = params[:rule]
-    task.hc_mask = 'NULL'
-  elsif params[:attackmode] == 'maskmode'
-    task.wl_id = 'NULL'
-    task.hc_rule = 'NULL'
-    task.hc_mask = params[:mask]
-  elsif params[:attackmode] == 'combinator'
-    task.wl_id = wordlist_list 
-    task.hc_rule = rule_list
-    task.hc_mask = 'NULL'
-  end
-  task.save
-
-  redirect to('/tasks/list')
-end
-
-get '/tasks/create' do
-  varWash(params)
-  @settings = Settings.first
-
-  # TODO present better error msg
-  flash[:warning] = 'You must define hashcat\'s binary path in global settings first.' if @settings && @settings.hcbinpath.nil?
-
-  @rules = []
-  # list wordlists that can be used
-  Dir.foreach('control/rules/') do |item|
-    next if item == '.' || item == '..'
-      @rules << item
-  end
-
-  @wordlists = Wordlists.all
-
-  haml :task_edit
-end
-
-post '/tasks/create' do
-  varWash(params)
-  settings = Settings.first
-  if settings && !settings.hcbinpath
-    flash[:error] = 'No hashcat binary path is defined in global settings.'
-    redirect to('/settings')
-  end
-
-  if !params[:name] || params[:name].empty?
-    flash[:error] = 'You must provide a name for your task!'
-    redirect to('/tasks/create')
-  end
-
-  @tasks = Tasks.all(name: params[:name])
-  unless @tasks.nil?
-    @tasks.each do |task|
-      if task.name == params[:name]
-        flash[:error] = 'Name already in use, pick another'
-        redirect to('/tasks/create')
-      end
-    end
-  end
-
-  wordlist = Wordlists.first(id: params[:wordlist])
-
-  # mask field cannot be empty
-  if params[:attackmode] == 'maskmode'
-    if !params[:mask] || params[:mask].empty?
-      flash[:error] = 'Mask field cannot be left empty'
-      redirect to('/tasks/create')
-    end
-  end
-
-  # must have two word lists
-  if params[:attackmode] == 'combinator'
-    wordlist_count = 0
-    wordlist_list = ''
-    rule_list = ''
-    @wordlists = Wordlists.all
-    @wordlists.each do |wordlist_check|
-      params.keys.each do |key|
-        if params[key] == 'on' && key == "combinator_wordlist_#{wordlist_check.id}"
-          if wordlist_list == ''
-            wordlist_list = wordlist_check.id.to_s + ','
-          else
-            wordlist_list = wordlist_list + wordlist_check.id.to_s
-          end
-          wordlist_count = wordlist_count + 1
-        end
-      end
-    end
-
-    if wordlist_count != 2
-      flash[:error] = 'You must specify at exactly 2 wordlists.'
-      redirect to('/tasks/create')
-    end
-
-    if params[:combinator_left_rule] && !params[:combinator_left_rule].empty? && params[:combinator_right_rule] && !params[:combinator_right_rule].empty?
-      rule_list = '--rule-left=' + params[:combinator_left_rule] + ' --rule-right=' + params[:combinator_right_rule]
-    elsif params[:combinator_left_rule] && !params[:combinator_left_rule].empty?
-      rule_list = '--rule-left=' + params[:combinator_left_rule]
-    elsif params[:combinator_right_rule] && !params[:combinator_right_rule].empty?
-      rule_list = '--rule-right=' + params[:combinator_right_rule]
-    else
-      rule_list = ''
-    end
-  end
-
-  task = Tasks.new
-  task.name = params[:name]
-
-  task.hc_attackmode = params[:attackmode]
-
-  if params[:attackmode] == 'dictionary'
-    task.wl_id = wordlist.id
-    task.hc_rule = params[:rule]
-  elsif params[:attackmode] == 'maskmode'
-    task.hc_mask = params[:mask]
-  elsif params[:attackmode] == 'combinator'
-    task.wl_id = wordlist_list
-    task.hc_rule = rule_list
-  end
-  task.save
-
-  flash[:success] = "Task #{task.name} successfully created."
-
-  redirect to('/tasks/list')
-end
+## Moved to routes/tasks.rb
 
 ############################
