+++ routes/init.rb
@@ -1,0 +1,4 @@
+# encoding: utf-8
+require_relative 'customers'
+require_relative 'accounts'
+#require_relative 'main'
